Extract scanner class
